// packages/entities/entities-plugins/src/components/free-form/plugins/datakit/types.ts
export interface DatakitUIData {
  /**
   * UI nodes that store data like input/output fields, positions, and other metadata.
   */
  nodes?: UINode[]
  groups?: UIGroup[]
}
/**
 * The phase of the node in the request/response cycle.
 *
 * Note: Phases of implicit nodes are hardcoded and cannot be changed.
 */
export type NodePhase = 'request' | 'response'

/************************************************
 *             Editor global store              *
 ************************************************/

export type NodeId = `node:${number}`
export type EdgeId = `edge:${number}`
export type FieldId = `field:${number}`
export type GroupId = `${NodeId}:${BranchName}`
export type GroupName = `${NodeName}:${BranchName}`

export type IdConnection = NodeId | `${NodeId}.${FieldId}`

export interface NodeField {
  id: FieldId
  name: FieldName
}

export type UINode = {
  name: NodeName
  phase: NodePhase
  position: XYPosition
  fields: { input?: FieldName[], output?: FieldName[] }
  expanded: { input?: boolean, output?: boolean }
  /** hidden in UI */
  hidden?: boolean
}

export interface NodeInstance {
  id: NodeId
  type: NodeType
  name: NodeName
  phase: NodePhase
  position: XYPosition
  expanded: { input?: boolean, output?: boolean }
  fields: { input: NodeField[], output: NodeField[] }
  /** hidden in UI */
  hidden?: boolean
  config?: Record<string, unknown>
}
// ...
export interface UIGroup {
  name: GroupName
  position: XYPosition
}
// ...
export type DatakitPluginData = FreeFormPluginData<DatakitConfig, DatakitUIData>